+++ crates/alembic-cli/examples/applied_ops_adapter.rs
@@ -9,6 +9,8 @@ alembic_external_main!(AppliedOpsAdapter::default());
 pub struct AppliedOpsAdapter {}
 
 impl ExternalAdapter for AppliedOpsAdapter {
+    type Error = anyhow::Error;
+
     fn setup(&mut self, _configuration: &serde_yaml::Value) -> Result<()> {
         Ok(())
     }