+++ crates/alembic-cli/examples/minimal_external_adapter.rs
@@ -11,6 +11,8 @@ pub struct MinimalAdapter {}
 impl MinimalAdapter {}
 
 impl ExternalAdapter for MinimalAdapter {
+    type Error = anyhow::Error;
+
     fn setup(&mut self, _configuration: &serde_yaml::Value) -> Result<()> {
         Ok(())
     }