+++ crates/alembic-cli/examples/preview_error_adapter.rs
@@ -13,6 +13,8 @@ pub struct PreviewErrorAdapter {}
 impl PreviewErrorAdapter {}
 
 impl ExternalAdapter for PreviewErrorAdapter {
+    type Error = anyhow::Error;
+
     fn setup(&mut self, _configuration: &serde_yaml::Value) -> Result<()> {
         Ok(())
     }